fix(chat): add context menu to chat header
